Added two tests 1. Log in with invalid credentials( check that 'Sign in' is available, check that 'Error' pop-up appears, check that 'Sign Out' button is not present
2. Sign in with invalid data ( enter to username, e-mail,pass/ click on Sign in button) and check error pop-up for all fields
