+++ ilinskaProj/src/main/java/pages/LoginPage.java
@@ -1,6 +1,7 @@
 package pages;
 
 import org.junit.Assert;
+import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
@@ -12,6 +13,8 @@ public class LoginPage<errorMessage> extends ParentPage {
     private WebElement inputPassword;
     @FindBy(xpath = ".//button[text()='Sign In']")
     private WebElement buttonSignIn;
+    @FindBy(xpath = ".//button[text()='Sign Out']")
+    private WebElement buttonSignOut;
     @FindBy(xpath = ".//div[@class='alert alert-danger text-center' and text()='Invalid username / password']")
     private WebElement errorSignIN;
     @FindBy(xpath = ".//input[@placeholder='Pick a username']")
@@ -51,7 +54,19 @@ public class LoginPage<errorMessage> extends ParentPage {
 
     }
 
-    public void enterLoginSignUp(String login) {
+    public void isButtonSignOutPresent() {
+        try {
+            webDriver.findElement(By.xpath(".//button[text()='Sign Out']")).isDisplayed();
+            logger.info("Button 'Sign Out is present");
+        } catch (Exception e) {
+            logger.error("Button Sign Out is not present"+ e);
+            Assert.fail("Button Sign Out is not present");
+        }
+    }
+
+
+    public void enterLoginSignUp(String login)
+    {
         enterTextToElement(inputLOgin, login);
     }
 
@@ -70,20 +85,13 @@ public class LoginPage<errorMessage> extends ParentPage {
         clickOnElement(buttonSignIn);
 
     }
-//    public void isButtonSignOutPresent(){
-//        if(isElementPresent(buttonSignOut)){
-//            logger.info("Button is present");
-//        } else {
-//            logger.info("Please try again!");
-//        }
-//
-//    }
+
 
     public void clickOnOurApp() {
         clickOnElement(OurApp);
     }
 
-    public boolean isPopupDisplay(String message) {
+    public boolean isPopupDisplay() {
         return isElementPresent(buttonSignIn);
 
     }
@@ -92,16 +100,17 @@ public class LoginPage<errorMessage> extends ParentPage {
         enterTextToElement(inputPass, password);
     }
 
-    public boolean errorPopUpUserName(String message) {
+    public boolean errorPopUpUserName() {
         return isElementPresent(errorMessage);
 
     }
-    public boolean errorPopupEmail(String message){
+    public boolean errorPopupEmail(){
         return  isElementPresent(errorEmailMessage);
 
     }
 
-    public boolean errorPopupPass(String message){
-        return isElementPresent(errorPassword);
+    public boolean errorPopupPass(){
+
+    return isElementPresent(errorPassword);
     }
 }
